pratical_1.1.c
C Program to Find Pair with Given Sum in an Array

This C program takes an array of integers and a target sum x. It checks whether there are two elements in the array whose sum is exactly equal to x. If such a pair exists, it returns a confirmation message, otherwise, it indicates that no such pair is found.